backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
transformers==4.28.1
huggingface-hub==0.20.3
codecarbon==2.1.4
torch==1.13.1
pylint==2.17.0
python-dotenv==1.0.1
requests==2.31.0
numpy==1.25.2
pandas==2.1.4